Support ?embed mode during SSR too
This wasn't happening during development because the local server
doesn't do SSR like the production one does (bad idea IMO, having a
local server that does the same as production would be great).

I noticed the header, footer and sidebar were flashing for a little bit
on the Memfault app, and I thought this might be because of SSR, and how
`EmbedModeProvider` doesn't support it.

With this change, `EmbedModeProvider` gets the search params from the
React router instead of from `document`, which should work both during
SSR as well as in the browser.