MFLT-3556: Mimic main website dir structure in /embed
I'm trying to reuse the links in the web application, which we currently
use as `src` to an `iframe`, to add a link so that users can read the
docs on `docs.memfault.com` as well.

It soon became apparent that having a different URL structure for
embed mode and normal mode was a bad idea.

The main website look like this:

- `/docs/embedded/*.mdx`
- `/docs/android/*.mdx`

However embed mode looked like this:

- `/embed/*.mdx`

So now they are both namespaced the same way, with `android` and
`embedded`.

I've also added redirects to our Netlify config so that this can be
deployed without orchestration.